Follow up of #2892
